+++ src/main/java/net/lugo/lightoverlay/util/ColorHelper.java
@@ -10,16 +10,7 @@ public class ColorHelper {
         return lightLevel >= threshold ? ModConfig.validColor : ModConfig.invalidColor;
     }
 
-    public static float[] getOverlayColorFloats(int lightLevel, int threshold) {
-        Color color = getOverlayColor(lightLevel, threshold);
-        float r = color.getRed() / 255f;
-        float g = color.getGreen() / 255f;
-        float b = color.getBlue() / 255f;
-
-        return new float[]{r, g, b};
-    }
-
-    public static float[] getOverlayColorFloats(int lightLevel, ClientLevel level) {
-        return getOverlayColorFloats(lightLevel, ModConfig.lightLevelThresholdForDimension(level));
+    public static Color getOverlayColor(int lightLevel, ClientLevel level) {
+        return getOverlayColor(lightLevel, ModConfig.lightLevelThresholdForDimension(level));
     }
 }